requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-32
icu=57.1=vc14_0
jpeg=8d=vc14_2
libpng=1.6.22=vc14_0
openssl=1.0.2j=vc14_0
pip=9.0.1=py35_1
pyqt=5.6.0=py35_1
python=3.5.2=0
pywin32=220=py35_1
qt=5.6.2=vc14_0
setuptools=27.2.0=py35_1
sip=4.18=py35_0
sqlalchemy=1.1.4=py35_0
vs2015_runtime=14.0.25123=0
wheel=0.29.0=py35_0
zlib=1.2.8=vc14_3
